% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZ_test.R
\name{MZ_test}
\alias{MZ_test}
\title{MZ_test}
\usage{
MZ_test(y, tau, H, yhat, B, l)
}
\arguments{
\item{y}{the realizations, a vector}

\item{tau}{vector of quantile levels, for which forecasts are provided}

\item{H}{maximum forecast horizon}

\item{yhat}{the forecasts, a list of length length(tau), which contains length(y) x H matrices with forecasts for a specific quantile for horizons 1 to H corresponding to the realizations (e.g. row 1 contains 1- up to H-period-ahead forecasts for time 1, where the first value in the realizations vector is also for time 1)}

\item{B}{number of bootstrap draws for the moving block bootstrap}

\item{l}{block length for the moving block bootstrap}
}
\value{
a list with two elements;
\itemize{
\item \code{resultsMZ} main results of the test, a named one matrix with one row containing five numbers:
\itemize{
\item \code{Stat} value of the MZ test statistic
\item \verb{90\%/95\%/99\%} 0.9/0.95/0.99 quantile of bootstrap distribution
\item \code{p-value} p-value of the test
}
\item \code{tables} additional information for each horizon-quantile combination, a list with three data frames:
\itemize{
\item \code{cont} individual contributions to test statistic
\item \code{alphahat} MZ regression intercepts
\item \code{betahat} MZ regression slopes
}
}
}
\description{
executes the MZ test for a given evaluation sample
}
