% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZ_test_augmented.R
\name{MZ_test_augmented}
\alias{MZ_test_augmented}
\title{MZ_test_augmented}
\usage{
MZ_test_augmented(y, tau, H, yhat, B, l, adreg)
}
\arguments{
\item{y}{the realizations, a vector}

\item{tau}{vector of quantile levels, for which forecasts are provided}

\item{H}{maximum forecast horizon}

\item{yhat}{the forecasts, a list of length length(tau), which contains length(y) x H matrices with forecasts for a specific quantile for horizons 1 to H corresponding to the realizations (e.g. row 1 contains 1- up to H-period-ahead forecasts for time 1, where the first value in the realizations vector is also for time 1)}

\item{B}{number of bootstrap draws for the moving block bootstrap}

\item{l}{block length for the moving block bootstrap}

\item{adreg}{additional regressors,  a (length(y) + H - 1) X A matrix with additional regressors to include in the MZ regression, A is the number of additional regressors; the columns contain the time series of additional regressors (known at the time of forecasting, i.e. from the forecaster's information set); the time period corresponding to a realization and forecasts at time t is t-h, i.e. this series has to start H periods prior to y and yhat and to end one period earlier}
}
\value{
a list with two elements;
\itemize{
\item \code{resultsMZ} main results of the test, a named one matrix with one row containing five numbers:
\itemize{
\item \code{Stat} value of the MZ test statistic
\item \verb{90\%/95\%/99\%} 0.9/0.95/0.99 quantile of bootstrap distribution
\item \code{p-value} p-value of the test
}
\item \code{tables} additional information for each horizon-quantile combination, a list with 3+A data frames:
\itemize{
\item \code{cont} individual contributions to test statistic
\item \code{alphahat} MZ regression intercepts
\item \code{betahat} MZ regression slopes
\item data frames 4 to 3+A.: regression coefficients of additional regressors
}
}
}
\description{
executes the augmented MZ test for a given evaluation sample and given additional regressors
}
