% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZ_test_multi.R
\name{MZ_test_multi}
\alias{MZ_test_multi}
\title{MZ_test_multi}
\usage{
MZ_test_multi(y, tau, H, yhat, B, l)
}
\arguments{
\item{y}{the realizations, a list of length G (the number of time series) with vectors with realizations (call their lengths T)}

\item{tau}{vector of quantile levels, for which forecasts are provided}

\item{H}{maximum forecast horizon}

\item{yhat}{the forecasts, a list of length G containing lists of length length(tau), which contain T x H matrices with forecasts for a specific quantile for horizons 1 to H corresponding to the realizations (e.g. row 1 contains 1- up to H-period-ahead forecasts for time 1, where the first value in the realizations vector is also for time 1)}

\item{B}{number of bootstrap draws for the moving block bootstrap}

\item{l}{block length for the moving block bootstrap}
}
\value{
a list with two elements;
\itemize{
\item \code{resultsMZ} results of the test, a named one matrix with one row containing five numbers:
\itemize{
\item \code{Stat} value of the MZ test statistic
\item \verb{90\%/95\%/99\%} 0.9/0.95/0.99 quantile of bootstrap distribution
\item \code{p-value} p-value of the test
}
}
}
\description{
executes the MZ test for multiple time series for a given evaluation sample
}
