% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZ_test_statistic.R
\name{MZ_test_statistic}
\alias{MZ_test_statistic}
\title{MZ_test_statistic}
\usage{
MZ_test_statistic(y, tau, H, yhat)
}
\arguments{
\item{y}{the realizations, a vector}

\item{tau}{vector of quantile levels, for which forecasts are provided}

\item{H}{maximum forecast horizon}

\item{yhat}{the forecasts, a list of length length(tau), which contains length(y) x H matrices with forecasts for a specific quantile for horizons 1 to H corresponding to the realizations (e.g. row 1 contains 1- up to H-period-ahead forecasts for time 1, where the first value in the realizations vector is also for time 1)}
}
\value{
a list with three elements;
\itemize{
\item \code{Uhat_MZ} value of the MZ test statistic
\item additional information for each horizon-quantile combination, a list with three data frames:
\itemize{
\item \code{cont} individual contributions to test statistic
\item \code{alphahat} MZ regression intercepts
\item \code{betahat} MZ regression slopes
}
\item \code{mhat} vector of the individual mhats (see paper for their definition)
}
}
\description{
computes the MZ test statistic from an evaluation sample
}
