% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZ_test_statistic_augmented.R
\name{MZ_test_statistic_augmented}
\alias{MZ_test_statistic_augmented}
\title{MZ_test_statistic_augmented}
\usage{
MZ_test_statistic_augmented(y, tau, H, yhat, adregmat)
}
\arguments{
\item{y}{the realizations, a vector}

\item{tau}{vector of quantile levels, for which forecasts are provided}

\item{H}{maximum forecast horizon}

\item{yhat}{the forecasts, a list of length length(tau), which contains length(y) x H matrices with forecasts for a specific quantile for horizons 1 to H corresponding to the realizations (e.g. row 1 contains 1- up to H-period-ahead forecasts for time 1, where the first value in the realizations vector is also for time 1)}

\item{adregmat}{additional regressors, a list of length A (the number of additional regressors), with length(y) x H matrices with additional regressors for horizons 1 to H corresponding to the realizations and forecasts; i.e. column hcount contains in every row the corresponding regressors for the MZ regression using the same row in yhat and y; this matrix structure is necessary in the bootstrap, where this function is used repeatedly}
}
\value{
a list with three elements;
\itemize{
\item \code{Uhat_MZ} value of the MZ test statistic
\item additional information for each horizon-quantile combination, a list with 3+A data frames:
\itemize{
\item \code{cont} individual contributions to test statistic
\item \code{alphahat} MZ regression intercepts
\item \code{betahat} MZ regression slopes
\item data frames 4 to 3+A.: regression coefficients of additional regressors
}
\item \code{mhat} vector of the individual mhats (see paper for their definition)
}
}
\description{
computes the test statistic of the augmented MZ test from an evaluation sample
}
