% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZ_test_statistic_multi.R
\name{MZ_test_statistic_multi}
\alias{MZ_test_statistic_multi}
\title{MZ_test_statistic_multi}
\usage{
MZ_test_statistic_multi(y, tau, H, yhat)
}
\arguments{
\item{y}{the realizations, a list of length G (the number of time series) with vectors with realizations (call their lengths T)}

\item{tau}{vector of quantile levels, for which forecasts are provided}

\item{H}{maximum forecast horizon}

\item{yhat}{the forecasts, a list of length G containing lists of length length(tau), which contain T x H matrices with forecasts for a specific quantile for horizons 1 to H corresponding to the realizations (e.g. row 1 contains 1- up to H-period-ahead forecasts for time 1, where the first value in the realizations vector is also for time 1)}
}
\value{
a list with two elements;
\itemize{
\item \code{Uhat_MZ} value of the MZ test statistic
\item \code{mhat} vector of the individual mhats (see paper for their definition)
}
}
\description{
computes the MZ test statistic for multiple time series from an evaluation sample
}
